fixed entrypoint.sh file exist problem

+++ v1/apache/entrypoint.sh
@@ -24,7 +24,10 @@ echo >&2 "MasterLab has been successfully copied to $PWD"
 
 chown -R $MASTERLAB_USER:$MASTERLAB_USER $MASTERLAB_MOUNT_POINT
 
-mkdir -p /var/spool/cron/crontabs
+if [ ! -d /var/spool/cron/crontabs ]; then
+    mkdir -p /var/spool/cron/crontabs
+fi
+
 cat > /var/spool/cron/crontabs/$MASTERLAB_USER <<EOF
 0,30 22-23 * * * php -f $MASTERLAB_MOUNT_POINT/app/server/timer/project.php
 55 23 * * * php -f $MASTERLAB_MOUNT_POINT/app/server/timer/projectDayReport.php
@@ -41,38 +44,41 @@ else
     MASTERLAB_DOMAIN_CUSTOM="ServerName  ${MASTERLAB_DOMAIN}"
 fi
 
-# modify apache config
-mv /etc/apache2/sites-enabled/000-default.conf /etc/apache2/sites-enabled/000-default.conf.bak
-cat > /etc/apache2/sites-enabled/masterlab.conf <<EOF
-<VirtualHost *:80>
-    DocumentRoot /var/www/html/app/public
-    # 这里修改成你自己的域名
-    ${MASTERLAB_DOMAIN_CUSTOM}
-    <Directory />
-        Options Indexes FollowSymLinks
-        AllowOverride All
-        Allow from All
-    </Directory>
-    <Directory /var/www/html/app/public>
-        Options  Indexes FollowSymLinks
-        AllowOverride All
-        Order allow,deny
-        Allow from All
-    </Directory>
+if [ ! -f /etc/apache2/sites-enabled/masterlab.conf ]; then
+    # modify apache config
+    mv /etc/apache2/sites-enabled/000-default.conf /etc/apache2/sites-enabled/000-default.conf.bak
+    cat > /etc/apache2/sites-enabled/masterlab.conf <<EOF
+    <VirtualHost *:80>
+        DocumentRoot /var/www/html/app/public
+        # 这里修改成你自己的域名
+        ${MASTERLAB_DOMAIN_CUSTOM}
+        <Directory />
+            Options Indexes FollowSymLinks
+            AllowOverride All
+            Allow from All
+        </Directory>
+        <Directory /var/www/html/app/public>
+            Options  Indexes FollowSymLinks
+            AllowOverride All
+            Order allow,deny
+            Allow from All
+        </Directory>
 
-    Alias /attachment /var/www/html/app/storage/attachment
-    <Directory /var/www/html/app/storage/attachment>
-        Options Indexes FollowSymLinks
-        AllowOverride All
-        Order allow,deny
-        Allow from all
-    </Directory>
+        Alias /attachment /var/www/html/app/storage/attachment
+        <Directory /var/www/html/app/storage/attachment>
+            Options Indexes FollowSymLinks
+            AllowOverride All
+            Order allow,deny
+            Allow from all
+        </Directory>
 
-    ErrorLog \${APACHE_LOG_DIR}/error.log
-    CustomLog \${APACHE_LOG_DIR}/access.log combined
-</VirtualHost>
-EOF
+        ErrorLog \${APACHE_LOG_DIR}/error.log
+        CustomLog \${APACHE_LOG_DIR}/access.log combined
+    </VirtualHost>
+    EOF
 
-echo >&2 "Complete! MasterLab-configure Initializing finished"
+    echo >&2 "Complete! MasterLab-configure Initializing finished"
+fi
+
 
 exec "$@"
